+++ applications/fishing-map/src/types/index.ts
@@ -19,6 +19,8 @@ export type WorkspaceStateProperty =
   | 'timebarGraph'
   | 'bivariate'
   | 'version'
+  | 'activityCategory'
+
 export type WorkspaceParam =
   | WorkspaceViewportParam
   | WorkspaceTimeRangeParam
@@ -31,16 +33,19 @@ export type WorkspaceAnalysis = {
   sourceId: string
   bounds?: [number, number, number, number]
 }
+export type WorkspaceActivityCategory = 'fishing' | 'presence'
+
 export type WorkspaceState = {
   query?: string
+  version?: string
   sidebarOpen?: boolean
   analysis?: WorkspaceAnalysis
   dataviewInstances?: Partial<UrlDataviewInstance[]>
   timebarVisualisation?: TimebarVisualisations
   timebarEvents?: TimebarEvents
   timebarGraph?: TimebarGraphs
-  bivariate?: boolean
-  version?: string
+  bivariate?: boolean // TODO move this between layers saving layer indexes
+  activityCategory?: WorkspaceActivityCategory
 }
 export type QueryParams = Partial<WorkspaceViewport> & Partial<WorkspaceTimeRange> & WorkspaceState