+++ applications/vessel-history/src/features/i18n/i18n.ts
@@ -38,4 +38,8 @@ i18n
     },
   })
 
+const t = i18n.t.bind(i18n)
+
+export { t }
+
 export default i18n