+++ packages/dataviews-client/src/resolve-dataviews-generators.ts
@@ -15,6 +15,7 @@ import {
   GeneratorDataviewConfig,
   Generators,
   Group,
+  COLOR_RAMP_DEFAULT_NUM_STEPS,
 } from '@globalfishingwatch/layer-composer'
 import type {
   ColorRampsIds,
@@ -183,12 +184,16 @@ export function getGeneratorConfig(
         ...generator,
         ...environmentalConfig,
       }
-      const tilesAPI = dataset?.endpoints?.find(
-        (endpoint) => endpoint.id === EndpointId.FourwingsTiles
-      )
-      const breaksAPI = dataset?.endpoints?.find(
-        (endpoint) => endpoint.id === EndpointId.FourwingsBreaks
-      )
+
+      // TODO use this instead of hardcoded version of the api endpoint in layer composer
+      // const { url: tilesAPI, dataset: heatmapDataset } = resolveDataviewDatasetResource(
+      //   dataview,
+      //   DatasetTypes.Fourwings
+      // )
+      // const breaksAPI = heatmapDataset?.endpoints?.find(
+      //   (endpoint) => endpoint.id === EndpointId.FourwingsBreaks
+      // )?.pathTemplate
+
       const visible = generator.sublayers?.some(({ visible }) => visible === true)
       const { extentStart, extentEnd } = getDatasetsExtent(dataview.datasets)
 
@@ -197,8 +202,8 @@ export function getGeneratorConfig(
         visible,
         debug,
         debugLabels: debug,
-        tilesAPI,
-        breaksAPI,
+        // tilesAPI,
+        // breaksAPI,
         ...(extentStart && { datasetsStart: extentStart }),
         ...(extentEnd && { datasetsEnd: extentEnd }),
         staticStart: timeRange?.start,
@@ -245,7 +250,7 @@ export function getGeneratorConfig(
             (dataset.configuration as EnviromentalDatasetConfiguration)?.propertyToIncludeRange ||
             {}
           const rampScale = scaleLinear().range([min, max]).domain([0, 1])
-          const numSteps = 8
+          const numSteps = COLOR_RAMP_DEFAULT_NUM_STEPS
           const steps = [...Array(numSteps)]
             .map((_, i) => parseFloat((i / (numSteps - 1))?.toFixed(2)))
             .map((value) => parseFloat((rampScale(value) as number)?.toFixed(3)))
@@ -284,7 +289,8 @@ export function getDataviewsGeneratorConfigs(
   // Collect heatmap animated generators and filter them out from main dataview list
   const dataviewsFiltered = dataviews.filter((d) => {
     const isActivityDataview =
-      d.category === DataviewCategory.Activity && d.config?.type === Generators.Type.HeatmapAnimated
+      (d.category === DataviewCategory.Fishing || d.category === DataviewCategory.Presence) &&
+      d.config?.type === Generators.Type.HeatmapAnimated
     if (isActivityDataview) {
       activityDataviews.push(d)
     }