+++ packages/layer-composer/src/index.ts
@@ -6,6 +6,7 @@ export {
   DEFAULT_HEATMAP_INTERVALS,
   TEMPORALGRID_SOURCE_LAYER,
   CONFIG_BY_INTERVAL,
+  COLOR_RAMP_DEFAULT_NUM_STEPS,
 } from './generators'
 export * from './types'
 export { frameToDate, quantizeOffsetToDate } from './generators/heatmap/util/time-chunks'