+++ applications/dataviews-editor/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @globalfishingwatchapp/dataviews-editor
 
+## 1.1.42
+
+### Patch Changes
+
+- Updated dependencies [67fdfec5]
+  - @globalfishingwatch/dataviews-client@6.5.1
+  - @globalfishingwatch/layer-composer@6.0.3
+
+## 1.1.41
+
+### Patch Changes
+
+- Updated dependencies [cfae2b4b]
+- Updated dependencies [cfae2b4b]
+  - @globalfishingwatch/layer-composer@6.0.2
+  - @globalfishingwatch/dataviews-client@6.5.0
+
+## 1.1.40
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/dataviews-client@6.4.2
+  - @globalfishingwatch/layer-composer@6.0.1
+
 ## 1.1.39
 
 ### Patch Changes