+++ packages/api-types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @globalfishingwatch/api-types
 
+## 4.2.0
+
+### Minor Changes
+
+- cfae2b4b: new dataview categories
+
+## 4.1.3
+
+### Patch Changes
+
+- 2f7db18f: Fishing map/color cycling
+
+## 4.1.2
+
+### Patch Changes
+
+- a5339163: fix tiles api
+
 ## 4.1.1
 
 ### Patch Changes