+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @globalfishingwatch/react-hooks
 
+## 6.1.4
+
+### Patch Changes
+
+- Updated dependencies [67fdfec5]
+  - @globalfishingwatch/dataviews-client@6.5.1
+  - @globalfishingwatch/layer-composer@6.0.3
+  - @globalfishingwatch/ui-components@3.0.5
+
+## 6.1.3
+
+### Patch Changes
+
+- Updated dependencies [cfae2b4b]
+- Updated dependencies [cfae2b4b]
+- Updated dependencies [cfae2b4b]
+  - @globalfishingwatch/ui-components@3.0.4
+  - @globalfishingwatch/layer-composer@6.0.2
+  - @globalfishingwatch/dataviews-client@6.5.0
+
+## 6.1.2
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/dataviews-client@6.4.2
+  - @globalfishingwatch/layer-composer@6.0.1
+  - @globalfishingwatch/ui-components@3.0.2
+
 ## 6.1.1
 
 ### Patch Changes